// src/store/modules/app/types.d.ts
// mutation
export enum MutationTypes_APP {
  SET_COUNTER = 'app/set-counter',
  SET_LANGUAGE = 'app/set-language',
  SET_COLLAPSE = 'app/set-collapse',
  SET_USERNAME = 'app/set-username',
}

// action
export enum ActionTypes_APP {
  SET_LANGUAGE = 'app/set-language',
}

// getter
export enum GetterTypes_APP {
  COUNTER = 'app/counter',
  LANGUAGE = 'app/language',
  COLLAPSE = 'app/collapse',
  USERNAME = 'app/username',
}